deduplicate Encapsulated, move to parser.rs
